refactor capabilities sdk
