+++ capabilities/client/pkg/discovery/cluster_schema.go
