+++ capabilities/client/pkg/discovery/cluster_test.go
