+++ capabilities/client/pkg/discovery/fake.go
