+++ capabilities/client/pkg/discovery/generate.go
