+++ capabilities/client/pkg/discovery/tkg/resource_test.go
