+++ capabilities/controller/go.mod
@@ -3,10 +3,8 @@ module github.com/vmware-tanzu/tanzu-framework/capabilities/controller
 go 1.17
 
 replace (
-	github.com/vmware-tanzu/tanzu-framework/apis/cli => ../../apis/cli
-	github.com/vmware-tanzu/tanzu-framework/apis/config => ../../apis/config
 	github.com/vmware-tanzu/tanzu-framework/apis/run => ../../apis/run
-	github.com/vmware-tanzu/tanzu-framework/capabilities/discovery => ../discovery
+	github.com/vmware-tanzu/tanzu-framework/capabilities/client => ../client
 	github.com/vmware-tanzu/tanzu-framework/cli/runtime => ../../cli/runtime
 	sigs.k8s.io/cluster-api => sigs.k8s.io/cluster-api v1.1.5
 )
@@ -16,7 +14,7 @@ require (
 	github.com/onsi/ginkgo v1.16.5
 	github.com/onsi/gomega v1.19.0
 	github.com/vmware-tanzu/tanzu-framework/apis/run v0.0.0-00010101000000-000000000000
-	github.com/vmware-tanzu/tanzu-framework/capabilities/discovery v0.0.0-00010101000000-000000000000
+	github.com/vmware-tanzu/tanzu-framework/capabilities/client v0.0.0-00010101000000-000000000000
 	github.com/vmware-tanzu/tanzu-framework/cli/runtime v0.0.0-00010101000000-000000000000
 	k8s.io/api v0.23.5
 	k8s.io/apimachinery v0.23.5
