+++ cmd/cli/plugin/secret/main.go
@@ -9,7 +9,7 @@ import (
 	"github.com/aunum/log"
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 
-	capdiscovery "github.com/vmware-tanzu/tanzu-framework/capabilities/discovery"
+	capdiscovery "github.com/vmware-tanzu/tanzu-framework/capabilities/client/pkg/discovery"
 	cliapi "github.com/vmware-tanzu/tanzu-framework/cli/runtime/apis/cli/v1alpha1"
 	"github.com/vmware-tanzu/tanzu-framework/cli/runtime/plugin"
 	"github.com/vmware-tanzu/tanzu-framework/packageclients/pkg/kappclient"
