+++ featuregates/controller/Dockerfile
@@ -16,7 +16,7 @@ COPY featuregates/controller/go.sum featuregates/controller/go.sum
 COPY featuregates/client featuregates/client
 COPY apis/ apis/
 COPY cli/runtime cli/runtime
-COPY capabilities/discovery capabilities/discovery
+COPY capabilities/client capabilities/client
 
 WORKDIR featuregates/controller
 RUN  go mod download
