+++ pkg/v1/tkr/Dockerfile
@@ -17,12 +17,9 @@ COPY apis/ apis/
 COPY cli/runtime cli/runtime
 COPY packageclients/ packageclients/
 COPY tkg/ tkg/
-<<<<<<< HEAD
 COPY cli/core cli/core
-=======
-COPY capabilities/discovery capabilities/discovery
+COPY capabilities/client capabilities/client
 
->>>>>>> c1537f44 (fix docker builds of other components)
 RUN  go mod download
 # cache deps before building and copying source so that we don't need to re-download as much
 # and so that source changes don't invalidate our downloaded layer
