+++ pkg/v2/tkr/webhook/infra-machine/Dockerfile
@@ -18,7 +18,7 @@ COPY cli/runtime cli/runtime
 COPY packageclients/ packageclients/
 COPY tkg/ tkg/
 COPY cli/core cli/core
-COPY capabilities/discovery capabilities/discovery
+COPY capabilities/client capabilities/client
 
 RUN  go mod download
 # cache deps before building and copying source so that we don't need to re-download as much
